# packages/front/.openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/login":{"post":{"operationId":"login","summary":"ログイン","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/api/auth/logout":{"post":{"operationId":"logout","summary":"ログアウト","parameters":[],"responses":{"201":{"description":""}}}},"/api/auth":{"post":{"operationId":"authorize","summary":"認証","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/users":{"get":{"operationId":"getUsers","summary":"全ユーザー取得","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"operationId":"addUser","summary":"ユーザー追加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/users/{id}":{"patch":{"operationId":"updateUser","summary":"ユーザー編集","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/tasks":{"get":{"operationId":"getTasks","summary":"タスク一覧取得","parameters":[{"name":"userId","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}},"post":{"operationId":"addTask","summary":"タスク追加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/tasks/{id}":{"patch":{"operationId":"updateTask","summary":"タスク修正","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"delete":{"operationId":"deleteTask","summary":"タスク削除","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/tasks/{taskId}/assigns":{"post":{"operationId":"addTaskAssigner","summary":"タスク受領者追加","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"number"}}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskAssign"}}}}}}}},"/api/tasks/{taskId}/assigners/{assignUserId}":{"delete":{"operationId":"deleteTaskAssigner","summary":"タスク受領者削除","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"number"}},{"name":"assignUserId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/tasks/{taskId}/status":{"patch":{"operationId":"updateTaskStatus","summary":"タスクステータス更新","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskStatusDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatus"}}}}}}}},"info":{"title":"ShafukuAPI","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"test"},"password":{"type":"string","example":"test"}},"required":["username","password"]},"User":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"username":{"type":"string"},"authority":{"type":"number"}},"required":["name","authority"]},"CreateUserDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"password":{"type":"string"}},"required":["user","password"]},"UpdateUserDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"password":{"type":"string"}}},"TaskStatus":{"type":"object","properties":{"id":{"type":"number"},"taskId":{"type":"number"},"status":{"type":"number"}},"required":["taskId","status"]},"Task":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"contents":{"type":"string"},"rewards":{"type":"number"},"orderer":{"type":"string"},"startDate":{"type":"string"},"endDate":{"type":"string"},"numOfRecruit":{"type":"number"},"lastStatus":{"$ref":"#/components/schemas/TaskStatus"},"assigners":{"type":"array","items":{"$ref":"#/components/schemas/User"}}},"required":["title","contents","rewards","orderer","endDate","numOfRecruit"]},"TaskAssign":{"type":"object","properties":{"id":{"type":"number"},"taskId":{"type":"number"},"userId":{"type":"number"},"task":{"$ref":"#/components/schemas/Task"},"user":{"$ref":"#/components/schemas/User"}},"required":["taskId","userId"]},"UpdateTaskStatusDto":{"type":"object","properties":{"status":{"type":"number"}},"required":["status"]}}}}